requirements.txt
Django==4.1.5
django_crispy_forms==2.0
pandas==1.5.2
